[lint] s/app/android/ in namespace of menu.item.showAsAction
Lint error message:
Should use android:showAsAction when not using the appcompat library.

Description:
When using the appcompat library, menu resources should refer to the
showAsAction (or actionViewClass, or actionProviderClass) in the
app: namespace, not the android: namespace.  Similarly, when not using
the appcompat library, you should be using the android:showAsAction
(or actionViewClass, or actionProviderClass) attribute.
Issue id: AppCompatResource